+++ clown_sort/files/pdf_file.py
@@ -2,22 +2,28 @@
 Wrapper for PDF files.
 """
 import io
-import logging
+import os
+from pathlib import Path
 from typing import Optional
 
 from PIL import Image
-from pypdf import PdfReader
+from pypdf import PdfReader, PdfWriter
 from pypdf.errors import DependencyError, EmptyFileError
 from rich.console import Console
 from rich.panel import Panel
+from rich.text import Text
 
 from clown_sort.config import Config, check_for_pymupdf, log_optional_module_warning
 from clown_sort.files.image_file import ImageFile
 from clown_sort.files.sortable_file import SortableFile
-from clown_sort.util.constants import MIN_PDF_SIZE_TO_LOG_PROGRESS_TO_STDERR
+from clown_sort.lib.page_range import PageRange
+from clown_sort.util.constants import MIN_PDF_SIZE_TO_LOG_PROGRESS_TO_STDERR, PDF_ERRORS
+from clown_sort.util.filesystem_helper import create_dir_if_it_does_not_exist, insert_suffix_before_extension
 from clown_sort.util.logging import log
-from clown_sort.util.rich_helper import WARNING, console, stderr_console
+from clown_sort.util.rich_helper import WARNING, attention_getting_panel, console, mild_warning, stderr_console
+from clown_sort.util.string_helper import exception_str
 
+DEFAULT_PDF_ERRORS_DIR = Path.cwd().joinpath(PDF_ERRORS)
 MAX_DISPLAY_HEIGHT = 600
 SCALE_FACTOR = 0.4
 
@@ -25,12 +31,13 @@ SCALE_FACTOR = 0.4
 class PdfFile(SortableFile):
     is_presentable_in_popup = None
 
-    def extracted_text(self) -> Optional[str]:
+    def extracted_text(self, page_range: Optional[PageRange] = None) -> Optional[str]:
         """Use Tesseract to OCR the text in the image, which is returned as a string."""
         if self.text_extraction_attempted:
             return self._extracted_text
 
         log.debug(f"Extracting text from '{self.file_path}'...")
+        self.page_numbers_of_errors = []
         extracted_pages = []
 
         try:
@@ -39,6 +46,10 @@ class PdfFile(SortableFile):
             log.debug(f"PDF Page count: {page_count}")
 
             for page_number, page in enumerate(pdf_reader.pages, start=1):
+                if page_range and not page_range.in_range(page_number):
+                    self._log_to_stderr(f"Skipping page {page_number}...")
+                    continue
+
                 self._log_to_stderr(f"Parsing page {page_number}...")
                 page_buffer = Console(file=io.StringIO())
                 page_buffer.print(Panel(f"PAGE {page_number}", padding=(0, 15), expand=False))
@@ -54,20 +65,24 @@ class PdfFile(SortableFile):
                         image_obj = Image.open(io.BytesIO(image.data))
                         image_text = ImageFile.extract_text(image_obj, f"{self.file_path} ({image_name})")
                         page_buffer.print((image_text or '').strip())
-                except NotImplementedError as e:
-                    stderr_console.print(f"WARNING: {type(e).__name__}: {e} while parsing embedded image {image_number} on page {page_number}...")
+                except (OSError, NotImplementedError, TypeError, ValueError) as e:
+                    error_str = exception_str(e)
+                    msg = f"{error_str} while parsing embedded image {image_number} on page {page_number}..."
+                    mild_warning(msg)
 
+                    # Dump an error PDF and encourage user to report to pypdf team.
                     if 'JBIG2Decode' not in str(e):
                         stderr_console.print_exception()
-                except (OSError, TypeError, ValueError) as e:
-                    stderr_console.print(f"WARNING: {type(e).__name__}: {e} while parsing embedded image {image_number} on page {page_number}...")
-                    stderr_console.print_exception()
+
+                        if page_number not in self.page_numbers_of_errors:
+                            self._handle_extraction_error(page_number, error_str)
+                            self.page_numbers_of_errors.append(page_number)
 
                 page_text = page_buffer.file.getvalue()
                 extracted_pages.append(page_text)
                 log.debug(page_text)
 
-                if Config.print_when_parsed:
+                if Config.print_as_parsed:
                     print(f"{page_text}")
         except DependencyError:
             log_optional_module_warning('pdf')
@@ -104,6 +119,40 @@ class PdfFile(SortableFile):
 
         return page.get_pixmap(matrix=zoom_matrix, clip= clip, alpha=False).tobytes()
 
+    def extract_page_range(
+            self,
+            page_range: PageRange,
+            destination_dir: Optional[Path] = None,
+            extra_file_suffix: Optional[str] = None
+        ) -> Path:
+        """Extract a range of pages to a new PDF file (or 1 page if last_page_number not provided.)"""
+        destination_dir = destination_dir or DEFAULT_PDF_ERRORS_DIR
+        create_dir_if_it_does_not_exist(destination_dir)
+
+        if extra_file_suffix is None:
+            file_suffix = page_range.file_suffix()
+        else:
+            file_suffix = f"{page_range.file_suffix()}__{extra_file_suffix}"
+
+        extracted_pages_pdf_basename = insert_suffix_before_extension(self.file_path, file_suffix).name
+        extracted_pages_pdf_path = destination_dir.joinpath(extracted_pages_pdf_basename)
+        stderr_console.print(f"Extracting {page_range.file_suffix()} from '{self.file_path}' to '{extracted_pages_pdf_path}'...")
+        pdf_writer = PdfWriter()
+
+        with open(self.file_path, 'rb') as source_pdf:
+            pdf_writer.append(fileobj=source_pdf, pages=page_range.to_tuple())
+
+        if SortableFile.confirm_file_overwrite(extracted_pages_pdf_path):
+            with open(extracted_pages_pdf_path, 'wb') as extracted_pages_pdf:
+                pdf_writer.write(extracted_pages_pdf)
+
+        stderr_console.print(f"Wrote new PDF '{extracted_pages_pdf_path}'.")
+        return extracted_pages_pdf_path
+
+    def print_extracted_text(self, page_range: Optional[PageRange] = None) -> None:
+        console.print(self._filename_panel())
+        console.print(self.extracted_text(page_range=page_range))
+
     def _can_be_presented_in_popup(self) -> bool:
         if type(self).is_presentable_in_popup is None:
             type(self).is_presentable_in_popup = check_for_pymupdf()
@@ -121,6 +170,30 @@ class PdfFile(SortableFile):
             return
 
         stderr_console.print(msg)
+
+    def _handle_extraction_error(self, page_number: int, error_msg: str) -> None:
+        """Rip the offending page to a new file and suggest that user report bug to PyPDF."""
+        if 'pdf_errors_dir' in dir(Config):
+            destination_dir = Config.pdf_errors_dir
+        else:
+            destination_dir = DEFAULT_PDF_ERRORS_DIR
+
+        extracted_file = self.extract_page_range(PageRange(str(page_number)), destination_dir, error_msg)
+
+        blink_txt = Text('', style='bright_white')
+        blink_txt.append("An error (", style='blink color(154)').append(error_msg, style='color(11) blink')
+        blink_txt.append(') ', style='blink color(154)')
+        blink_txt.append("was encountered while processing a PDF file.\n\n", style='blink color(154)')
+
+        txt = Text(f"The error was of a type such that it probably came from a bug in ", style='bright_white')
+        txt.append('PyPDF', style='underline bright_green').append('. It was encountered processing the file ')
+        txt.append(str(self.file_path), style='file').append('. You should see a stack trace above this box.\n\n')
+
+        txt.append('The offending page will be extracted to ', style='bright_white')
+        txt.append(str(extracted_file), style='file').append('.\n\n')
+        txt.append(f"Please visit 'https://github.com/py-pdf/pypdf/issues' to report a bug. ", style='bold')
+        txt.append(f"Providing the devs with the extracted page and the stack trace help improve pypdf.")
+        stderr_console.print(attention_getting_panel(blink_txt + txt, title='PyPDF Error'))
 
     def __repr__(self) -> str:
         return f"PdfFile('{self.file_path}')"