+++ clown_sort/util/constants.py
@@ -10,6 +10,7 @@ MAC_DEFAULT_SCREENSHOTS_DIR = Path.home().joinpath('Pictures', 'Screenshots')
 
 # Miscellaneous strings
 CRYPTO = 'crypto'
+PDF_ERRORS = 'PDF Errors'
 
 
 ### Environment variables